src/microservices/events/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
kafka-python==2.0.2
pydantic==2.6.3
loguru==0.7.2